Major Refactor
* convert from big.Int to []byte

This is more general and also moves the problem of zero prefix handling
into the library itself where it can't be accidentally missed by a user.
See block.go for how that was handled.

* remove interfaces

This library had a number of interfaces that were there with the
intention that maybe, one day, someone might want to override certain
aspects. However that day never came and it was unnecessarily
complicated. Eventually if we want that flexibility back it can happen
under the go generics regime instead.

* harden cobra usage

This removes places where previously it was just panic'ing and uses the
facilities in cobra to ensure the right number of args are passed.

* Tests!

Turns out we never had tests for this. Now we do. Huzzah!

* Docs!

Also didn't really have godoc docs. Now we at least have some!

* Easier creation with NewIBF

This function now just takes a size and seed. This is what most people
want. For those that want control over the individual hashers
NewIBFWithHash is available.

* Removed version subcommand

Until this is wired in automatically it was going to be a nuisance.
Later if there is time we can make the build process do the needful.